% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dev_ranking.R
\name{create_dev_ranking}
\alias{create_dev_ranking}
\title{create_dev_ranking}
\usage{
create_dev_ranking(new_connect_rsc_users, new_connect_rsc_content)
}
\arguments{
\item{new_connect_rsc_users}{}

\item{new_connect_rsc_content}{}
}
\description{
create_dev_ranking
}
