% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_migration.R
\name{detect_migration}
\alias{detect_migration}
\title{detect_migration}
\usage{
detect_migration(data, threshold = 0.5)
}
\arguments{
\item{data}{}

\item{threshold}{}
}
\description{
detect_migration
}
