% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_migration.R
\name{detect_migration_and_testapps}
\alias{detect_migration_and_testapps}
\title{detect_migration_and_testapps}
\usage{
detect_migration_and_testapps(data, threshold = 0.5)
}
\arguments{
\item{data}{}

\item{threshold}{}
}
\description{
detect_migration_and_testapps
}
