% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_users_content.R
\name{fetch_users_content}
\alias{fetch_users_content}
\title{fetch_users_content}
\usage{
fetch_users_content(connect_server = server, connect_api_key = api_key)
}
\arguments{
\item{connect_server}{}

\item{connect_api_key}{}
}
\description{
fetch_users_content
}
